Extract DSL classes into DSL namespace
